requirements.txt
absl-py==0.13.0
anyio==3.2.0
appnope==0.1.2
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==3.3.0
blis==0.4.1
cached-property==1.5.2
cachetools==4.2.2
catalogue==1.0.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
decorator==5.0.9
defusedxml==0.7.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz
entrypoints==0.3
findspark==1.4.2
Flask==2.0.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
# h5py==3.1.0
idna==2.10
importlib-metadata==4.5.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.8.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
kafka-python==2.0.2
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
# murmurhash==1.0.5
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.4.0
# numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.5
pandocfilters==1.4.3
parso==0.8.2
pathy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
plac==1.1.3
# preshed==3.0.5
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.4
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
pyspark==3.1.2
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.1.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
Send2Trash==1.7.1
six==1.15.0
sklearn==0.0
smart-open==3.0.0
sniffio==1.2.0
spacy==2.2.4
spacy-legacy==3.0.6
srsly==1.0.5
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
textblob==0.15.3
thinc==7.4.0
threadpoolctl==2.2.0
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
tweepy==3.10.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.5
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
